requirements.txt
djangorestframework==3.10.3
urllib3==1.25.3
requests==2.22.0
django==2.2.7
xmltodict==0.12.0
